requirements.txt
jsonschema>=4.18.0
pyyaml
